+++ backend/src/main/java/com/trend_now/backend/config/quartz/QuartzJobConfig.java
@@ -1,28 +1,21 @@
-package com.trend_now.backend.config;
+package com.trend_now.backend.config.quartz;
 
 import com.trend_now.backend.board.application.SignalKeywordJob;
-import com.trend_now.backend.board.application.SignalKeywordJobListener;
 import com.trend_now.backend.post.application.PostLikesSyncDbJob;
-import com.trend_now.backend.post.application.PostLikesSyncDbJobListener;
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.quartz.JobBuilder;
 import org.quartz.JobDataMap;
 import org.quartz.JobDetail;
-import org.quartz.Scheduler;
-import org.quartz.SchedulerException;
 import org.quartz.SimpleScheduleBuilder;
 import org.quartz.Trigger;
 import org.quartz.TriggerBuilder;
-import org.quartz.impl.StdSchedulerFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @RequiredArgsConstructor
-@ConditionalOnProperty(name = "quartzScheduler.enabled", havingValue = "true", matchIfMissing = false)
-public class QuartzSchedulerConfig {
+public class QuartzJobConfig {
 
     private static final String SIGNAL_KEYWORD_GROUP = "SignalGroup";
     private static final String SIGNAL_KEYWORD_JOB = "SignalKeywordJob";
@@ -33,23 +26,27 @@ public class QuartzSchedulerConfig {
     private static final String POST_LIKES_SYNC_DB_JOB_TRIGGER = "PostLikesSyncDbJobTrigger";
     private static final int POST_LIKES_SYNC_DB_JOB_INTERVAL_SECONDS = 11;
 
-    private final Scheduler scheduler;
     private final ApplicationContext applicationContext;
 
-    @PostConstruct
-    public void scheduleTopKeywordJob() throws SchedulerException {
+    @Bean
+    public JobDetail signalKeywordJobDetail() {
         JobDataMap ctx = new JobDataMap();
         ctx.put("applicationContext", applicationContext);
 
-        JobDetail job = JobBuilder
+        return JobBuilder
                 .newJob(SignalKeywordJob.class)
                 .withIdentity(SIGNAL_KEYWORD_JOB, SIGNAL_KEYWORD_GROUP)
                 .withDescription("TOP10 검색어 순위 리스트 Job")
                 .setJobData(ctx)
+                .storeDurably()
                 .build();
+    }
 
-        Trigger trigger = TriggerBuilder
+    @Bean
+    public Trigger signalKeywordTrigger(JobDetail signalKeywordJobDetail) {
+        return TriggerBuilder
                 .newTrigger()
+                .forJob(signalKeywordJobDetail)
                 .withIdentity(SIGNAL_KEYWORD_TRIGGER, SIGNAL_KEYWORD_GROUP)
                 .withDescription("TOP10 검색어 순위 리스트 Trigger")
                 .startNow()
@@ -59,19 +56,27 @@ public class QuartzSchedulerConfig {
                                 .withIntervalInSeconds(SIGNAL_KEYWORD_SCHEDULER_INTERVAL_SECONDS)
                                 .repeatForever())
                 .build();
+    }
 
+    @Bean
+    public JobDetail postLikesSyncDbJobDetail() {
         JobDataMap postLikesCtx = new JobDataMap();
         postLikesCtx.put("applicationContext", applicationContext);
 
-        JobDetail postLikesJob = JobBuilder
+        return JobBuilder
                 .newJob(PostLikesSyncDbJob.class)
                 .withIdentity(POST_LIKES_SYNC_DB_JOB, POST_LIKES_SYNC_DB_JOB_GROUP)
                 .withDescription("DB 동기화 job")
                 .setJobData(postLikesCtx)
+                .storeDurably()
                 .build();
+    }
 
-        Trigger postLikesTrigger = TriggerBuilder
+    @Bean
+    public Trigger postLikesSyncDbTrigger(JobDetail postLikesSyncDbJobDetail) {
+        return TriggerBuilder
                 .newTrigger()
+                .forJob(postLikesSyncDbJobDetail)
                 .withIdentity(POST_LIKES_SYNC_DB_JOB_TRIGGER, POST_LIKES_SYNC_DB_JOB_GROUP)
                 .withDescription("DB 동기화 Trigger")
                 .startNow()
@@ -81,14 +86,5 @@ public class QuartzSchedulerConfig {
                                 .withIntervalInSeconds(POST_LIKES_SYNC_DB_JOB_INTERVAL_SECONDS)
                                 .repeatForever())
                 .build();
-
-        SignalKeywordJobListener signalKeywordJobListener = new SignalKeywordJobListener();
-        PostLikesSyncDbJobListener postLikesSyncDbJobListener = new PostLikesSyncDbJobListener();
-        scheduler.getListenerManager().addJobListener(signalKeywordJobListener);
-        scheduler.getListenerManager().addJobListener(postLikesSyncDbJobListener);
-
-        scheduler.start();
-        scheduler.scheduleJob(job, trigger);
-        scheduler.scheduleJob(postLikesJob, postLikesTrigger);
     }
 }
